run_self_improvement.py
#!/usr/bin/env python3
"""
TRUE ASI System - Self-Improvement Demo
========================================

Demonstrate the self-improvement engine capabilities.

Author: TRUE ASI System
Date: November 1, 2025
Version: 1.0.0
"""

import os
import sys
import asyncio
import logging
from pathlib import Path
from dotenv import load_dotenv

# Add src to path
sys.path.insert(0, str(Path(__file__).parent / 'src'))

# Import components
from self_improvement.self_improvement_engine import SelfImprovementEngine

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)


# Mock classes for demonstration
class MockLLMClient:
    """Mock LLM client for demonstration"""
    async def generate(self, prompt, model=None):
        return "Generated algorithm using advanced optimization techniques..."


class MockAWSIntegration:
    """Mock AWS integration for demonstration"""
    def __init__(self):
        import boto3
        self.s3 = boto3.client('s3')
        self.bucket = os.getenv('S3_BUCKET', 'asi-knowledge-base-898982995956')


async def main():
    """Main demonstration function"""
    print("="*70)
    print("TRUE ASI SYSTEM - SELF-IMPROVEMENT ENGINE DEMONSTRATION")
    print("="*70)
    print()
    
    # Load environment
    load_dotenv()
    
    # Initialize components
    llm_client = MockLLMClient()
    aws_integration = MockAWSIntegration()
    
    # Create self-improvement engine
    engine = SelfImprovementEngine(llm_client, aws_integration)
    
    print("🚀 Starting self-improvement session (3 cycles)...")
    print()
    
    # Run improvement session
    session_result = await engine.run_improvement_session(num_cycles=3)
    
    print()
    print("="*70)
    print("SESSION RESULTS")
    print("="*70)
    print(f"Cycles Completed: {session_result['num_cycles']}")
    print(f"Total Improvement: {session_result['final_performance']}")
    print(f"Duration: {session_result['started_at']} to {session_result['completed_at']}")
    print()
    
    # Generate and display report
    report = engine.generate_report()
    print(report)
    
    # Save report
    report_file = Path("SELF_IMPROVEMENT_REPORT.txt")
    report_file.write_text(report)
    print()
    print(f"✅ Report saved: {report_file}")
    
    print()
    print("="*70)
    print("SELF-IMPROVEMENT DEMONSTRATION COMPLETE!")
    print("="*70)


if __name__ == "__main__":
    asyncio.run(main())
