requirements.txt
# TRUE ASI System Dependencies

# Core
python>=3.11

# AI/ML
openai>=1.0.0
anthropic>=0.7.0

# AWS
boto3>=1.28.0
botocore>=1.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.1

# Async
asyncio>=3.4.3
aiohttp>=3.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
